% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation_data.R
\name{get_elevation_data}
\alias{get_elevation_data}
\title{Retrieve elevation data for US geographies}
\usage{
get_elevation_data(level, geoid, year = 2022, resolution = "500k", z = 8)
}
\arguments{
\item{level}{Character. One of "state", "county" or "tract"}

\item{geoid}{Character. The FIPS code of your geography.
Two digits for state, five for county, eleven for tract.}

\item{year}{Numeric. Year for the geographic boundaries.
Defaults to 2022.}

\item{resolution}{Character. Resolution for the cartographic
boundary file retrieved by tigris::. Defaults to "500K". Other options are "5m" (1:5 million) and "20m" (1:20  million).}

\item{z}{Resolution used by elevatr to retrieve elevation raster.
Defaults to 8.}
}
\value{
List containing: "map"
 (a map of your geography showing elevation);
 "elevation_mean" (elevation of your geography);
 "raster" (raster file returned by elevatr and
 cropped to your geographic polygon);
 and "elevation.df" (elevation data in a dataframe with
 coordinates in 'x' and 'y' columns).
}
\description{
Retrieve elevation data for US geographies
}
\examples{
\dontrun{
elev <- get_elevation_data(level = "state", geoid = "41")
print(paste("Mean elevation:",elev$elevation_mean,"meters"))
elev$map

elev <- get_elevation_data(level = "county", geoid = "24021", z = 10)
print(paste("Mean elevation:",elev$elevation_mean,"meters"))
elev$map

elev <- get_elevation_data(level = "tract", geoid = "24510200100", z = 8)
print(paste("Mean elevation:",elev$elevation_mean,"meters"))
elev$map
}
}
