% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation_data_batch.R
\name{get_elevation_data_batch}
\alias{get_elevation_data_batch}
\title{Retrieve elevation data for all substate geographies within a state}
\usage{
get_elevation_data_batch(
  level,
  state,
  county = NULL,
  year = 2022,
  resolution = "500k",
  z = 8
)
}
\arguments{
\item{level}{Character. One of "state", "county" or "tract"}

\item{state}{Character. Two-digit FIPS code for a state.}

\item{county}{Character. Optional. If level = "tract", county can be set to a
three-digit county fips code (e.g., "001") to limit only to tracts within than county.}

\item{year}{Numeric. Year for the geographic boundaries.
Defaults to 2022.}

\item{resolution}{Character. Resolution for the cartographic
boundary file retrieved by tigris::. Defaults to "500K". Other options are "5m" (1:5 million) and "20m" (1:20  million).}

\item{z}{Resolution used by elevatr to retrieve elevation raster.
Defaults to 8.}
}
\value{
dataframe of substate geographies with elevation data
}
\description{
Retrieve elevation data for all substate geographies within a state
}
\examples{
\dontrun{
elev_data <- get_elevation_data_batch(level = "county", state = "24") # All counties in Maryland
elev_data <- get_elevation_data_batch(level = "tract", state = "24") # All tracts in Maryland
elev_data <- get_elevation_data_batch(level = "tract", state = "24", county = "001") # Tracts in Allegany County, Maryland
}
}
